requirement.txt
pytorch
Pillow
opencv-python
numpy